fix warnings
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@1044 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ examples/vehiclemmi/vehiclemmi_core.ss/src/VehicleMmiCppQt/entitydialog.cpp
@@ -132,7 +132,7 @@ namespace VehicleMmiCppQt
         {
             if(!pVehicle->Position()->Latitude().IsNull())
             {
-                float posLat = pVehicle->Position()->Latitude().GetVal();
+                double posLat = pVehicle->Position()->Latitude().GetVal();
                 QString qstrPosLat("");
                 qstrPosLat.setNum(posLat);
                 ui.lineEditPositionLat->setText(qstrPosLat);
@@ -140,7 +140,7 @@ namespace VehicleMmiCppQt
 
             if(!pVehicle->Position()->Longitude().IsNull())
             {
-                float posLong = pVehicle->Position()->Longitude().GetVal();
+                double posLong = pVehicle->Position()->Longitude().GetVal();
                 QString qstrPosLong("");
                 qstrPosLong.setNum(posLong);
                 ui.lineEditPositionLong->setText(qstrPosLong);
